+++ build.gradle
@@ -91,6 +91,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:${versions.gradle}"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${versions.kotlin}"
+        classpath "com.karumi:shot:5.13.0"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
