+++ showkase-screenshot-testing-shot/src/main/java/com/vinaygaba/showkase_screenshot_testing_shot/ShotShowkaseScreenshotTest.kt
@@ -1,0 +1,20 @@
+package com.vinaygaba.showkase_screenshot_testing_shot
+
+import android.graphics.Bitmap
+import com.airbnb.android.showkase.screenshot.testing.ShowkaseScreenshotTest
+import com.airbnb.android.showkase.screenshot.testing.ShowkaseScreenshotType
+import com.karumi.shot.ScreenshotTest
+
+abstract class ShotShowkaseScreenshotTest: ShowkaseScreenshotTest, ScreenshotTest {
+
+    override fun onScreenshot(
+        id: String,
+        name: String,
+        group: String,
+        styleName: String?,
+        screenshotType: ShowkaseScreenshotType,
+        screenshotBitmap: Bitmap
+    ) {
+        compareScreenshot(composeTestRule)
+    }
+}
